new commit

+++ Assets/Assets/Scripts/PlayerScripts/Spells/PoisonCircle.cs
@@ -4,7 +4,7 @@ public class PoisonCircleScript : PlayerSpellBaseScript
 {
     public override float Speed => GlobalVariables.Instance.poisonCircleSpeed;
     public override float Damage => GlobalVariables.Instance.playerAttackDamage * GlobalVariables.Instance.poisonCircleDamageMulti;
-    public override float Bounces => GlobalVariables.Instance.poisonCircleBounces;
+    public override int Bounces => GlobalVariables.Instance.poisonCircleBounces;
     public override float KnockbackForce => GlobalVariables.Instance.poisonCircleKnockbackForce;
     public override float CriticalChance => GlobalVariables.Instance.poisonCircleCriticalChance;
     public override float CriticalMultiplier => GlobalVariables.Instance.poisonCircleCriticalMultiplier;
