+++ Assets/Assets/Scripts/PlayerScripts/Spells/darkballScript.cs
@@ -5,7 +5,7 @@ public class darkballScript : PlayerSpellBaseScript
 {
     public override float Speed => GlobalVariables.Instance.darkballSpeed;
     public override float Damage => GlobalVariables.Instance.playerAttackDamage * GlobalVariables.Instance.darkballDamageMulti;
-    public override float Bounces => GlobalVariables.Instance.darkballBounces;
+    public override int Bounces => GlobalVariables.Instance.darkballBounces;
     public override float KnockbackForce => GlobalVariables.Instance.darkballKnockbackForce;
     public override float CriticalChance => GlobalVariables.Instance.darkballCriticalChance;
     public override float CriticalMultiplier => GlobalVariables.Instance.darkballCriticalMultiplier;
