+++ Assets/Assets/Scripts/PlayerScripts/Spells/fireballScript.cs
@@ -4,7 +4,7 @@ public class FireballScript : PlayerSpellBaseScript
 {
     public override float Speed => GlobalVariables.Instance.fireballSpeed;
     public override float Damage => GlobalVariables.Instance.playerAttackDamage * GlobalVariables.Instance.fireballDamage;
-    public override float Bounces => GlobalVariables.Instance.fireballBounces;
+    public override int Bounces => GlobalVariables.Instance.fireballBounces;
     public override float KnockbackForce => GlobalVariables.Instance.fireballKnockbackForce;
     public override float CriticalChance => GlobalVariables.Instance.fireballCriticalChance;
     public override float CriticalMultiplier => GlobalVariables.Instance.fireballCriticalMultiplier;
